www: Extract ProjectBuildersWidget component
